move test folder to root directory
