+++ __test__/queries/getFavoriteCourse.test.js
@@ -1,6 +1,6 @@
-const connection = require('../../src/database/config/connection');
-const dbBuild = require('../../src/database/config/build');
-const { getFavoriteCourse } = require('../../src/database/queries');
+const connection = require('../../server/src/database/config/connection');
+const dbBuild = require('../../server/src/database/config/build');
+const { getFavoriteCourse } = require('../../server/src/database/queries');
 
 describe('test for getFavoriteCourse query', () => {
   beforeAll(() => dbBuild());
