+++ package.json
@@ -10,7 +10,7 @@
     "lint-fix": "cd client && npm run lint-fix && cd ../server && npm run lint-fix",
     "server": "cd server && cross-env NODE_ENV=development npm run dev",
     "client": "cd client && npm start",
-    "test": "cd server && cross-env NODE_ENV=test npm test",
+    "test": "NODE_ENV=test jest --runInBand",
     "kill": "fuser -k 5000/tcp",
     "kill-run": "npm run kill && NODE_ENV=development npm run server",
     "heroku-postbuild": "NODE_ENV=production cd client && npm install && npm run build",
@@ -31,8 +31,10 @@
     "cross-env": "^7.0.2"
   },
   "devDependencies": {
+    "jest": "^26.0.1",
     "npm-run-all": "^4.1.5",
-    "pre-commit": "^1.2.2"
+    "pre-commit": "^1.2.2",
+    "supertest": "^4.0.2"
   },
   "pre-commit": [
     "test",
