+++ server/src/app.js
@@ -19,10 +19,8 @@ app.use(express.static(join(__dirname, '..', '..', 'client', 'build')));
 
 app.use('/api/v1', router);
 
-if (process.env.NODE_ENV === 'production') {
-  app.get('*', (req, res) => {
-    res.sendFile(join(__dirname, '..', '..', 'client', 'build', 'index.html'));
-  });
-}
+app.get('*', (req, res) => {
+  res.sendFile(join(__dirname, '..', '..', 'client', 'build', 'index.html'));
+});
 
 module.exports = app;
